refactor(kornia-py): move Tensor out of kornia_rs.cuda to top-level kornia_rs
Tensor mirrors kornia_tensor::Tensor, which isn't CUDA-specific — nesting it
under kornia_rs.cuda implied it was. Moved the pyclass to module="kornia_rs"
and register it at the top-level module (still #[cfg(feature="cuda")]-gated,
since the binding is currently device-only) instead of inside cuda::register().

kornia_rs.cuda.Tensor -> kornia_rs.Tensor; kornia_rs.cuda.CudaPreprocessor
(unaffected) still returns/consumes it. .pyi: the class stub moved to
__init__.pyi; cuda.pyi re-exports it (`from . import Tensor as Tensor`) for
CudaPreprocessor's signatures.

Verified: device suite 49 passed/4 skipped; full torch+TensorRT
re-verification 76 passed (re-run since this again touches the zero-copy
binding path); example runs; clippy clean.
